feat: integrate heightfield flood game into the main app + controls
Promote the flood prototype to the actual game and retire the old 2D-voxel code.

- src/bin/flood_demo.rs -> src/flood.rs as FloodPlugin; slim main.rs runs it,
  so `cargo run` launches the heightfield flood game.
- Delete the abandoned 2D modules (simulation, render, grid, ui, levels,
  persistence, undo) and the demo bins. camera.rs/textures.rs left unreferenced.
- UI panel: object weight buttons (200-5000kg), a Pour Water tool, and WAVE
  patterns (Flood/Sine/Random) modulating the source.
- Weight-scaled blocks: footprint + height scale with weight; dam height and
  obstacle footprint tied to size; collision spacing scales too.
- Orbit camera: right-drag orbit, middle-drag pan, scroll zoom (gentle).
- Placement indicator: gizmo box at the cursor sized to the selected weight.
- Pause: Space / top button freezes the water + object sim; camera, placement,
  and rendering keep running.

Foam and obstacle-raised water rendering were tried for overtopping and reverted
(both looked bad); clean flat water restored. Proper cresting needs a weir-model
rework, logged as next. PROGRESS.md updated (Session 17).
